+++ src/commands/index.named.ts
@@ -1,0 +1,3 @@
+export * from './runTranslateIt';
+export * from './clearTranslateIt';
+export * from './changeTargetLanguage';